+++ example/lib/main.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
-
-import 'package:logger/logger.dart';
-import 'package:logger_flutter/logger_flutter.dart';
+import 'package:logger_fork/logger_fork.dart';
+import 'package:logger_flutter_fork/logger_flutter_fork.dart';
 
 void main() {
-  runApp(MyApp());
+  LogConsole.init();
   log();
+  runApp(const MyApp());
 }
 
 var logger = Logger(
@@ -27,22 +26,60 @@ void log() {
   logger.e("Error! Something bad happened", "Test Error");
 
   loggerNoStack.v({"key": 5, "value": "something"});
+  loggerNoStack.wtf('What a terrible failure!');
 
-  Future.delayed(Duration(seconds: 5), log);
+  Future.delayed(const Duration(seconds: 5), log);
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: Scaffold(
-        body: LogConsoleOnShake(
-          dark: true,
-          child: Center(
-            child: Text("Shake Phone to open Console."),
-          ),
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({Key? key, required this.title}) : super(key: key);
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            LogConsoleOnShake(
+              child: const Text(
+                'Shake device to open the log console',
+              ),
+            ),
+            const Text(
+              'or tap on floating action button to see log console',
+            ),
+          ],
         ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () => LogConsole.open(context),
+        tooltip: 'info',
+        child: const Icon(Icons.info),
       ),
     );
   }