+++ lib/src/log_console.dart
@@ -1,4 +1,4 @@
-part of logger_flutter;
+part of logger_flutter_fork;
 
 ListQueue<OutputEvent> _outputEventBuffer = ListQueue();
 int _bufferSize = 20;
@@ -7,8 +7,10 @@ bool _initialized = false;
 class LogConsole extends StatefulWidget {
   final bool dark;
   final bool showCloseButton;
+  final bool showClearButton;
+  final void Function(String content)? onExport;
 
-  LogConsole({this.dark = false, this.showCloseButton = false})
+  LogConsole({this.dark = false, this.showCloseButton = false, this.showClearButton = true, this.onExport})
       : assert(_initialized, "Please call LogConsole.init() first.");
 
   static void init({int bufferSize = 20}) {
@@ -16,12 +18,29 @@ class LogConsole extends StatefulWidget {
 
     _bufferSize = bufferSize;
     _initialized = true;
-    Logger.addOutputListener((e) {
-      if (_outputEventBuffer.length == bufferSize) {
-        _outputEventBuffer.removeFirst();
-      }
-      _outputEventBuffer.add(e);
-    });
+
+    Logger.addOutputListener(LogOutputListener(
+      (e) {
+        if (_outputEventBuffer.length == bufferSize) {
+          _outputEventBuffer.removeFirst();
+        }
+        _outputEventBuffer.add(e);
+      },
+    ));
+  }
+
+  static void open(BuildContext context,
+      {bool dark = false, bool showCloseButton = false, bool showClearButton = true, void Function(String)? onExport}) {
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+          builder: (context) => LogConsole(
+                dark: dark,
+                showClearButton: showClearButton,
+                showCloseButton: showCloseButton,
+                onExport: onExport,
+              )),
+    );
   }
 
   @override
@@ -38,7 +57,7 @@ class RenderedEvent {
 }
 
 class _LogConsoleState extends State<LogConsole> {
-  OutputCallback _callback;
+  late LogOutputListener _callback;
 
   ListQueue<RenderedEvent> _renderedBuffer = ListQueue();
   List<RenderedEvent> _filteredBuffer = [];
@@ -57,21 +76,19 @@ class _LogConsoleState extends State<LogConsole> {
   void initState() {
     super.initState();
 
-    _callback = (e) {
+    _callback = LogOutputListener((e) {
       if (_renderedBuffer.length == _bufferSize) {
         _renderedBuffer.removeFirst();
       }
-
       _renderedBuffer.add(_renderEvent(e));
       _refreshFilter();
-    };
+    });
 
     Logger.addOutputListener(_callback);
 
     _scrollController.addListener(() {
       if (!_scrollListenerEnabled) return;
-      var scrolledToBottom = _scrollController.offset >=
-          _scrollController.position.maxScrollExtent;
+      var scrolledToBottom = _scrollController.offset >= _scrollController.position.maxScrollExtent;
       setState(() {
         _followBottom = scrolledToBottom;
       });
@@ -117,11 +134,11 @@ class _LogConsoleState extends State<LogConsole> {
       theme: widget.dark
           ? ThemeData(
               brightness: Brightness.dark,
-              accentColor: Colors.blueGrey,
+              colorScheme: ColorScheme.fromSwatch().copyWith(secondary: Colors.blueGrey),
             )
           : ThemeData(
               brightness: Brightness.light,
-              accentColor: Colors.lightBlueAccent,
+              colorScheme: ColorScheme.fromSwatch().copyWith(secondary: Colors.lightBlueAccent),
             ),
       home: Scaffold(
         body: SafeArea(
@@ -195,6 +212,37 @@ class _LogConsoleState extends State<LogConsole> {
             ),
           ),
           Spacer(),
+          if (widget.showClearButton)
+            IconButton(
+              icon: Icon(
+                Icons.delete,
+                color: Color.fromARGB(255, 254, 20, 3),
+              ),
+              onPressed: () {
+                _renderedBuffer.clear();
+                _outputEventBuffer.clear();
+                _refreshFilter();
+                setState(() {});
+              },
+            ),
+          IconButton(
+            icon: Icon(Icons.import_export),
+            onPressed: () {
+              var content = _renderedBuffer.map((e) => e.lowerCaseText).join();
+              if (widget.onExport != null) {
+                widget.onExport?.call(content);
+              } else {
+                if (kDebugMode) {
+                  log(
+                    'removed share functionality from logger package you need to add this manually',
+                    level: 24,
+                    error: UnimplementedError(),
+                  );
+                }
+                // Share.share(content);
+              }
+            },
+          ),
           IconButton(
             icon: Icon(Icons.add),
             onPressed: () {
@@ -241,7 +289,7 @@ class _LogConsoleState extends State<LogConsole> {
             ),
           ),
           SizedBox(width: 20),
-          DropdownButton(
+          DropdownButton<Level>(
             value: _filterLevel,
             items: [
               DropdownMenuItem(
@@ -270,7 +318,7 @@ class _LogConsoleState extends State<LogConsole> {
               )
             ],
             onChanged: (value) {
-              _filterLevel = value;
+              _filterLevel = value ?? Level.info;
               _refreshFilter();
             },
           )
@@ -319,7 +367,7 @@ class LogBar extends StatelessWidget {
   final bool dark;
   final Widget child;
 
-  LogBar({this.dark, this.child});
+  LogBar({required this.dark, required this.child});
 
   @override
   Widget build(BuildContext context) {
@@ -330,7 +378,7 @@ class LogBar extends StatelessWidget {
           boxShadow: [
             if (!dark)
               BoxShadow(
-                color: Colors.grey[400],
+                color: Colors.grey[400] ?? Colors.grey,
                 blurRadius: 3,
               ),
           ],
@@ -344,5 +392,16 @@ class LogBar extends StatelessWidget {
         ),
       ),
     );
+  }
+}
+
+class LogOutputListener extends LogOutput {
+  final void Function(OutputEvent) _listener;
+
+  LogOutputListener(this._listener);
+
+  @override
+  void output(OutputEvent event) {
+    _listener(event);
   }
 }