+++ lib/src/log_console_on_shake.dart
@@ -1,4 +1,4 @@
-part of logger_flutter;
+part of logger_flutter_fork;
 
 class LogConsoleOnShake extends StatefulWidget {
   final Widget child;
@@ -6,8 +6,8 @@ class LogConsoleOnShake extends StatefulWidget {
   final bool debugOnly;
 
   LogConsoleOnShake({
-    @required this.child,
-    this.dark,
+    required this.child,
+    this.dark = false,
     this.debugOnly = true,
   });
 
@@ -16,7 +16,7 @@ class LogConsoleOnShake extends StatefulWidget {
 }
 
 class _LogConsoleOnShakeState extends State<LogConsoleOnShake> {
-  ShakeDetector _detector;
+  late ShakeDetector _detector;
   bool _open = false;
 
   @override
@@ -44,14 +44,14 @@ class _LogConsoleOnShakeState extends State<LogConsoleOnShake> {
     _detector.startListening();
   }
 
-  _openLogConsole() async {
+  void _openLogConsole() async {
     if (_open) return;
 
     _open = true;
 
     var logConsole = LogConsole(
       showCloseButton: true,
-      dark: widget.dark ?? Theme.of(context).brightness == Brightness.dark,
+      dark: widget.dark,
     );
     PageRoute route;
     if (Platform.isIOS) {