+++ CHANGELOG.md
@@ -1,5 +1,16 @@
+## 1.3.0
+
+* removed share_plus plugin.
+    > if you need to share log handle it in your project.
+
+## 1.2.0
+
+* Null safe and working with logger_fork
+
 ## 0.7.1
-- Fixed error during build
+
+* Fixed error during build
 
 ## 0.7.0
-- First version
+
+* First version